Disable amd64 assembly on non-intel architectures
All builds would include an .S file which contains a pre-processor
directive that makes the file effectively empty on non-amd64
architectures. Objects built from assembly get an executable stack by
default. The file has a stanza to make the stack non-executable, but
on non-amd64 architectures this stanza is skipped together with the
rest of the file. Nevertheless, it seems that even the empty file
causes an executable stack to be created. To avoid the issue, do a
build with the assembler file explicitly skipped.